Fix PR feedback and add convenience method for reading static thresholds
